+++ recording/lib/edge_timestamp_saver.ex
@@ -17,7 +17,11 @@ defmodule Membrane.RTC.Engine.Endpoint.Recording.EdgeTimestampSaver do
                 """
               ]
 
-  @packets_interval 200
+  @packets_interval 100
+
+  @breakpoint 2 ** 31
+
+  defguardp in_order?(delta) when delta < -@breakpoint or (delta > 0 and delta < @breakpoint)
 
   @impl true
   def handle_init(_ctx, options) do
@@ -32,18 +36,32 @@ defmodule Membrane.RTC.Engine.Endpoint.Recording.EdgeTimestampSaver do
 
   @impl true
   def handle_buffer(_pad, %Buffer{} = buffer, ctx, %{first_buffer_timestamp: nil} = state) do
+    timestamp = buffer.metadata.rtp.timestamp
+
     Recording.Reporter.start_timestamp(
       state.reporter,
       track_id(ctx),
-      buffer.metadata.rtp.timestamp
+      timestamp
     )
 
     {[buffer: {:output, buffer}],
-     %{state | first_buffer_timestamp: buffer.metadata.rtp.timestamp}}
+     %{
+       state
+       | first_buffer_timestamp: timestamp,
+         last_buffer_timestamp: timestamp
+     }}
+  end
+
+  @impl true
+  def handle_buffer(_pad, %Buffer{} = buffer, _ctx, state)
+      when not in_order?(buffer.metadata.rtp.timestamp - state.last_buffer_timestamp) do
+    counter = min(state.counter + 1, @packets_interval - 1)
+    {[buffer: {:output, buffer}], %{state | counter: counter}}
   end
 
   @impl true
   def handle_buffer(_pad, %Buffer{} = buffer, ctx, state) do
+    timestamp = buffer.metadata.rtp.timestamp
     counter = rem(state.counter + 1, @packets_interval)
 
     if counter == 0,
@@ -51,11 +69,12 @@ defmodule Membrane.RTC.Engine.Endpoint.Recording.EdgeTimestampSaver do
         Recording.Reporter.end_timestamp(
           state.reporter,
           track_id(ctx),
-          buffer.metadata.rtp.timestamp
+          timestamp
         )
 
-    {[buffer: {:output, buffer}],
-     %{state | counter: counter, last_buffer_timestamp: buffer.metadata.rtp.timestamp}}
+    state = %{state | counter: counter, last_buffer_timestamp: timestamp}
+
+    {[buffer: {:output, buffer}], state}
   end
 
   @impl true