+++ recording/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## 0.2.0-dev
+* Fix setting last buffer timestamp [#394](https://github.com/jellyfish-dev/membrane_rtc_engine/pull/394)
 
 ## 0.1.0
 * Initial release [#356](https://github.com/jellyfish-dev/membrane_rtc_engine/pull/356)